requirements.txt
console-menu
pytube
requests
tqdm
ffmpeg-python
beautifulsoup4